[expo-asset] Move side effects into an .fx.ts file
